Initial commit
A simple example program that illustrates the use of oauth2 authentication,
for server-to-server interaction. In this case the application proves its
own identity to the API, with the use of an authentication key, without the
need for user consent. Delegation to domain-wide authority must be granted
to the service account. More information can be found on the following link:
https://developers.google.com/identity/protocols/OAuth2ServiceAccount